+++ .gitignore
@@ -5,3 +5,4 @@ compiled
 coverage
 .DS_Store
 next.d.ts
+legacy.d.ts
