+++ rollup.config.js
@@ -8,10 +8,10 @@ import uglify from 'rollup-plugin-uglify';
 const shared = {
   entry: `compiled/formik.js`,
   sourceMap: true,
-  external: ['react', 'prop-types'],
+  external: ['react', 'react-native'],
   globals: {
     react: 'React',
-    'prop-types': 'PropTypes',
+    'react-native': 'ReactNative',
   },
   exports: 'named',
 };
@@ -32,25 +32,74 @@ export default [
           process.env.NODE_ENV || 'development'
         ),
       }),
-      commonjs(),
+      resolve(),
+      commonjs({
+        include: /node_modules/,
+        namedExports: {
+          'node_modules/prop-types/index.js': ['object'],
+        },
+      }),
       sourceMaps(),
       process.env.NODE_ENV === 'production' && filesize(),
       process.env.NODE_ENV === 'production' && uglify(),
     ],
   }),
+
   Object.assign({}, shared, {
     targets: [
       { dest: 'dist/formik.es6.js', format: 'es' },
       { dest: 'dist/formik.js', format: 'cjs' },
     ],
-    plugins: [resolve(), sourceMaps()],
+    plugins: [
+      resolve(),
+      commonjs({
+        include: /node_modules/,
+        namedExports: {
+          'node_modules/prop-types/index.js': ['object'],
+        },
+      }),
+      ,
+      sourceMaps(),
+    ],
   }),
   Object.assign({}, shared, {
-    entry: `compiled/next.js`,
+    entry: `compiled/legacy.js`,
     targets: [
-      { dest: 'dist/next.es6.js', format: 'es' },
-      { dest: 'dist/next.js', format: 'cjs' },
+      { dest: 'dist/legacy.es6.js', format: 'es' },
+      { dest: 'dist/legacy.js', format: 'cjs' },
     ],
-    plugins: [resolve(), sourceMaps()],
+    plugins: [
+      resolve(),
+      commonjs({
+        include: /node_modules/,
+      }),
+      ,
+      sourceMaps(),
+    ],
+  }),
+  Object.assign({}, shared, {
+    moduleName: 'Formik',
+    format: 'umd',
+    entry: `compiled/legacy.js`,
+    dest:
+      process.env.NODE_ENV === 'production'
+        ? './dist/legacy.umd.min.js'
+        : './dist/legacy.umd.js',
+    plugins: [
+      resolve(),
+      replace({
+        exclude: 'node_modules/**',
+        'process.env.NODE_ENV': JSON.stringify(
+          process.env.NODE_ENV || 'development'
+        ),
+      }),
+      resolve(),
+      commonjs({
+        include: /node_modules/,
+      }),
+      sourceMaps(),
+      process.env.NODE_ENV === 'production' && filesize(),
+      process.env.NODE_ENV === 'production' && uglify(),
+    ],
   }),
 ];
