+++ test/formik.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
-import { Formik, InjectedFormikProps } from '../src/formik';
+import { Formik, FormikConfig, FormikProps } from '../src/formik';
 import { mount, shallow } from 'enzyme';
 
 const Yup = require('yup');
@@ -22,7 +22,7 @@ interface Values {
   name: string;
 }
 
-const Form: React.SFC<InjectedFormikProps<Props, Values>> = ({
+const Form: React.SFC<FormikProps<Values>> = ({
   values,
   handleSubmit,
   handleChange,
@@ -62,18 +62,17 @@ const Form: React.SFC<InjectedFormikProps<Props, Values>> = ({
   );
 };
 
-const FormFactory = (options = {}) =>
-  Formik<Props, Values, Values>({
-    mapPropsToValues: ({ user }) => ({ ...user }),
-    handleSubmit: noop,
-    ...options,
-  })(Form);
+const BasicForm = (
+  <Formik
+    getInitialValues={{ name: 'jared' }}
+    handleSubmit={noop}
+    component={Form}
+  />
+);
 
-const BasicForm = FormFactory();
-
-describe('Formik', () => {
+describe('Formik Next', () => {
   it('should initialize Formik state and pass down props', () => {
-    const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+    const tree = shallow(BasicForm);
     expect(tree.find(Form).props().isSubmitting).toBe(false);
     expect(tree.find(Form).props().touched).toEqual({});
     expect(tree.find(Form).props().values).toEqual({ name: 'jared' });
@@ -85,7 +84,7 @@ describe('Formik', () => {
   describe('FormikHandlers', () => {
     describe('handleChange', () => {
       it('sets values state', async () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
 
         // Simulate a change event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('change', {
@@ -103,7 +102,7 @@ describe('Formik', () => {
       });
 
       it('updates values state via `name` instead of `id` attribute when both are present', async () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
 
         // Simulate a change event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('change', {
@@ -123,11 +122,15 @@ describe('Formik', () => {
 
       it('runs validations if validateOnChange is set to true', async () => {
         const validate = jest.fn(noop);
-        const ValidationForm = FormFactory({
-          validate,
-          validateOnChange: true,
-        });
-        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
+        const tree = shallow(
+          <Formik
+            getInitialValues={{ name: 'jared' }}
+            handleSubmit={noop}
+            component={Form}
+            validate={validate}
+            validateOnChange={true}
+          />
+        );
         tree.find(Form).dive().find('input').simulate('change', {
           persist: noop,
           target: {
@@ -138,10 +141,18 @@ describe('Formik', () => {
         expect(validate).toHaveBeenCalled();
       });
 
-      it('does NOT run validations by default or if validateOnChange is set to false', async () => {
+      it('does NOT run validations if validateOnChange is set to false', async () => {
         const validate = jest.fn(noop);
-        const ValidationForm = FormFactory({ validate });
-        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
+
+        const tree = shallow(
+          <Formik
+            getInitialValues={{ name: 'jared' }}
+            handleSubmit={noop}
+            component={Form}
+            validate={validate}
+            validateOnChange={false}
+          />
+        );
         tree.find(Form).dive().find('input').simulate('change', {
           persist: noop,
           target: {
@@ -155,7 +166,7 @@ describe('Formik', () => {
 
     describe('handleBlur', () => {
       it('sets touched state', () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
 
         // Simulate a blur event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('blur', {
@@ -168,7 +179,7 @@ describe('Formik', () => {
       });
 
       it('updates touched state via `name` instead of `id` attribute when both are present', () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
 
         // Simulate a blur event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('blur', {
@@ -181,31 +192,32 @@ describe('Formik', () => {
         expect(tree.update().state().touched).toEqual({ name: true });
       });
 
-      it('runs validations by default or if validateOnBlur is set to true ', async () => {
+      it('runs validations if validateOnBlur is set to true ', async () => {
         const validate = jest.fn(noop);
-        const ValidationForm = FormFactory({ validate });
-        const ValidationForm2 = FormFactory({ validate, validateOnBlur: true });
-        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
-        const tree2 = shallow(<ValidationForm2 user={{ name: 'jared' }} />);
+
+        const tree = shallow(
+          <Formik
+            getInitialValues={{ name: 'jared' }}
+            handleSubmit={noop}
+            component={Form}
+            validate={validate}
+            validateOnBlur={true}
+          />
+        );
+
         tree.find(Form).dive().find('input').simulate('blur', {
           persist: noop,
           target: {
             name: 'name',
           },
         });
-        tree2.find(Form).dive().find('input').simulate('blur', {
-          persist: noop,
-          target: {
-            name: 'name',
-          },
-        });
-        expect(validate).toHaveBeenCalledTimes(2);
+        expect(validate).toHaveBeenCalledTimes(1);
       });
     });
 
     describe('handleSubmit', () => {
       it('should call preventDefault()', () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
         const preventDefault = jest.fn();
         tree.find(Form).dive().find('form').simulate('submit', {
           preventDefault,
@@ -214,7 +226,7 @@ describe('Formik', () => {
       });
 
       it('should touch all fields', () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
         tree.find(Form).dive().find('form').simulate('submit', {
           preventDefault: noop,
         });
@@ -222,7 +234,7 @@ describe('Formik', () => {
       });
 
       it('should push submission state changes to child component', () => {
-        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+        const tree = shallow(BasicForm);
 
         expect(tree.find(Form).dive().find('#submitting')).toHaveLength(0);
 
@@ -233,33 +245,35 @@ describe('Formik', () => {
         expect(tree.find(Form).dive().find('#submitting')).toHaveLength(1);
       });
 
-      it('should correctly map form values to payload', () => {
-        interface Payload {
-          user: { name: string };
-        }
-        const CustomPayloadForm = Formik<Props, Values, Payload>({
-          mapPropsToValues: ({ user }) => ({ ...user }),
-          mapValuesToPayload: ({ name }) => ({ user: { name } }),
-          handleSubmit: payload => {
-            expect(payload).toEqual({ user: { name: 'jared' } });
-            expect(payload).not.toEqual({ name: 'jared' });
-          },
-        })(Form);
-        const tree = shallow(<CustomPayloadForm user={{ name: 'jared' }} />);
-        tree.find(Form).dive().find('form').simulate('submit', {
-          preventDefault: noop,
-        });
-      });
+      // it('should correctly map form values to payload', () => {
+      //   interface Payload {
+      //     user: { name: string };
+      //   }
+      //   const CustomPayloadForm = Formik<Props, Values, Payload>({
+      //     mapPropsToValues: ({ user }) => ({ ...user }),
+      //     mapValuesToPayload: ({ name }) => ({ user: { name } }),
+      //     handleSubmit: payload => {
+      //       expect(payload).toEqual({ user: { name: 'jared' } });
+      //       expect(payload).not.toEqual({ name: 'jared' });
+      //     },
+      //   })(Form);
+      //   const tree = shallow(<CustomPayloadForm user={{ name: 'jared' }} />);
+      //   tree.find(Form).dive().find('form').simulate('submit', {
+      //     preventDefault: noop,
+      //   });
+      // });
 
       describe('with validate (SYNC)', () => {
         it('should call validate if present', () => {
           const validate = jest.fn().mockReturnValue({});
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate,
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit: noop,
-          })(Form);
-          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={noop}
+              component={Form}
+              validate={validate}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -268,12 +282,14 @@ describe('Formik', () => {
 
         it('should submit the form if valid', () => {
           const handleSubmit = jest.fn();
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate: noop,
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit,
-          })(Form);
-          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={handleSubmit}
+              component={Form}
+              validate={noop}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -284,13 +300,14 @@ describe('Formik', () => {
           const validate = jest.fn().mockReturnValue({ name: 'Error!' });
           const handleSubmit = jest.fn();
 
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate,
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit,
-          })(Form);
-
-          const tree = shallow(<ValidateForm user={{ name: '' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={handleSubmit}
+              component={Form}
+              validate={validate}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -302,12 +319,15 @@ describe('Formik', () => {
       describe('with validate (ASYNC)', () => {
         it('should call validate if present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate,
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit: noop,
-          })(Form);
-          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
+
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={noop}
+              component={Form}
+              validate={validate}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -317,13 +337,14 @@ describe('Formik', () => {
         it('should submit the form if valid', async () => {
           const handleSubmit = jest.fn();
 
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate: () => Promise.resolve({}),
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit,
-          })(Form);
-
-          const tree = shallow(<ValidateForm user={{ name: '' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={handleSubmit}
+              component={Form}
+              validate={() => Promise.resolve({})}
+            />
+          );
           await tree.find(Form).props().submitForm();
 
           expect(handleSubmit).toHaveBeenCalled();
@@ -332,16 +353,17 @@ describe('Formik', () => {
         it('should not submit the form if invalid', async () => {
           const handleSubmit = jest.fn();
 
-          const ValidateForm = Formik<Props, Values, Values>({
-            validate: () =>
-              sleep(25).then(() => {
-                throw { name: 'error!' };
-              }),
-            mapPropsToValues: ({ user }) => ({ ...user }),
-            handleSubmit,
-          })(Form);
-
-          const tree = shallow(<ValidateForm user={{ name: '' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={handleSubmit}
+              component={Form}
+              validate={() =>
+                sleep(25).then(() => {
+                  throw { name: 'error!' };
+                })}
+            />
+          );
           await tree.find(Form).props().submitForm();
 
           expect(handleSubmit).not.toHaveBeenCalled();
@@ -351,8 +373,17 @@ describe('Formik', () => {
       describe('with validationSchema (ASYNC)', () => {
         it('should run validationSchema if present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-          const ValidateForm = FormFactory({ validationSchema: { validate } });
-          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={noop}
+              component={Form}
+              validate={validate}
+              validationSchema={{
+                validate,
+              }}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -361,12 +392,17 @@ describe('Formik', () => {
 
         it('should call validationSchema if it is a function and present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-          const ValidateForm = FormFactory({
-            validationSchema: () => ({
-              validate,
-            }),
-          });
-          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
+          const tree = shallow(
+            <Formik
+              getInitialValues={{ name: 'jared' }}
+              handleSubmit={noop}
+              component={Form}
+              validate={validate}
+              validationSchema={() => ({
+                validate,
+              })}
+            />
+          );
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -378,84 +414,117 @@ describe('Formik', () => {
 
   describe('FormikActions', () => {
     it('setValues sets values', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(tree.find(Form).dive().find('input').props().value).toEqual('ian');
     });
 
     it('setValues should run validations when validateOnChange is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-        validateOnChange: true,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnChange={true}
+        />
+      );
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(validate).toHaveBeenCalled();
     });
 
-    it('setValues should NOT run validations when validateOnChange is false or undefined', () => {
+    it('setValues should NOT run validations when validateOnChange is false', () => {
       const validate = jest.fn();
-      const ValidateOnBlurForm = FormFactory({ validate });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnChange={false}
+        />
+      );
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setFieldValue sets value by key', () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(tree.find(Form).dive().find('input').props().value).toEqual('ian');
     });
 
     it('setFieldValue should run validations when validateOnChange is true', () => {
       const validate = jest.fn().mockReturnValue({});
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-        validateOnChange: true,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnChange={true}
+        />
+      );
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(validate).toHaveBeenCalled();
     });
 
-    it('setFieldValue should NOT run validations when validateOnChange is false or undefined', () => {
+    it('setFieldValue should NOT run validations when validateOnChange is false', () => {
       const validate = jest.fn();
-      const ValidateOnBlurForm = FormFactory({ validate });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnChange={false}
+        />
+      );
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setTouched sets touched', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setTouched({ name: true });
       expect(tree.find(Form).props().touched).toEqual({ name: true });
     });
 
-    it('setTouched should run validations by default, or when validateOnBlur is true', async () => {
+    it('setTouched should NOT run validations by default', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
-      tree.find(Form).props().setTouched({ name: true });
-      expect(validate).toHaveBeenCalled();
-    });
-
-    it('setTouched should NOT run validations when validateOnBlur is false', () => {
-      const validate = jest.fn();
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-        validateOnBlur: false,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+        />
+      );
       tree.find(Form).props().setTouched({ name: true });
       expect(validate).not.toHaveBeenCalled();
     });
 
+    it('setTouched should run validations when validateOnBlur is true', () => {
+      const validate = jest.fn();
+
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnBlur={true}
+        />
+      );
+      tree.find(Form).props().setTouched({ name: true });
+      expect(validate).toHaveBeenCalled();
+    });
+
     it('setFieldTouched sets touched by key', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setFieldTouched('name', true);
       expect(tree.find(Form).props().touched).toEqual({ name: true });
       expect(tree.find(Form).props().dirty).toBe(true);
@@ -466,28 +535,37 @@ describe('Formik', () => {
 
     it('setFieldTouched should run validations when validateOnBlur is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-        validateOnBlur: true,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnBlur={true}
+        />
+      );
       tree.find(Form).props().setFieldTouched('name', true);
       expect(validate).toHaveBeenCalled();
     });
 
     it('setFieldTouched should NOT run validations when validateOnBlur is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const ValidateOnBlurForm = FormFactory({
-        validate,
-        validateOnBlur: false,
-      });
-      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          validate={validate}
+          validateOnBlur={false}
+        />
+      );
       tree.find(Form).props().setFieldTouched('name', true);
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setErrors sets error object', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setErrors({ name: 'Required' });
       expect(tree.find(Form).dive().find('#feedback').text()).toEqual(
         'Required'
@@ -495,7 +573,7 @@ describe('Formik', () => {
     });
 
     it('setFieldError sets error by key', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).props().setFieldError('name', 'Required');
       expect(tree.find(Form).dive().find('#feedback').text()).toEqual(
         'Required'
@@ -503,7 +581,7 @@ describe('Formik', () => {
     });
 
     it('setStatus sets status object', async () => {
-      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.find(Form).dive().find('#statusButton').simulate('click');
       expect(tree.find(Form).dive().find('#statusMessage')).toHaveLength(1);
     });
@@ -511,52 +589,73 @@ describe('Formik', () => {
 
   describe('FormikComputedProps', () => {
     it('should compute dirty as soon as any input is touched', () => {
-      const ValidForm = FormFactory();
-      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       expect(tree.find(Form).props().dirty).toBe(false);
       tree.setState({ touched: { name: true } });
       expect(tree.find(Form).props().dirty).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and returns true', () => {
-      const InvalidForm = FormFactory({ isInitialValid: props => true });
-      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          isInitialValid={props => true}
+        />
+      );
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and returns false', () => {
-      const InvalidForm = FormFactory({ isInitialValid: props => false });
-      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          isInitialValid={props => false}
+        />
+      );
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if isInitialValid boolean is present and set to true', () => {
-      const InvalidForm = FormFactory({ isInitialValid: true });
-      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          isInitialValid={true}
+        />
+      );
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and set to false', () => {
-      const InvalidForm = FormFactory({ isInitialValid: false });
-      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
+      const tree = shallow(
+        <Formik
+          getInitialValues={{ name: 'jared' }}
+          handleSubmit={noop}
+          component={Form}
+          isInitialValid={false}
+        />
+      );
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if the form is dirty and there are errors', () => {
-      const ValidForm = FormFactory();
-      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.setState({ touched: { name: true }, errors: { name: 'Required!' } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if the form is dirty and there are not errors', () => {
-      const ValidForm = FormFactory();
-      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
+      const tree = shallow(BasicForm);
       tree.setState({ touched: { name: true } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(true);
